Remove check-ui Player component
Instead of using `Player` from `check-ui`, which relies on `react-player`, we detect whether something is a Youtube video. If it is a Youtube video, we embed it as an iframe and disable the playback speed options in the check user interface. If it is a native video, we use the `<video>` tag and control that via the check user interface as normal. This is going to let us remove `Player` from `check-ui`, reducing the bundle size by about 200k.