Remove react-color library and use HTML input instead
By using `<input type="color">` instead of the `react-color` library, we can remove 130kb of dependencies. This commit refactors the `ColorPicker` component to use the basic browser color picker. Since styling options are limited for the color picker button, we keep the old UI button and overlay an invisible version of the browser color picker button, resulting in 100% identical behavior as we had before, but with less JS cost.

Reference: CHECK-2815