Update check-ui
This updates us to a new check-ui that reduces the size of check-ui by 300kb (!) by including the *distribution* version of Leaflet rather than the *development* version. They use the dev version in all their examples and don't actually tell you how to include the distribution version which is... a choice!