Fixing build and unit tests
Probably this is what was causing the error 137 on CI. Locally, I noticed that the build was failing because the new components `TextField` and `TextArea` are not used by any other component in the application. We didn't have a way to tell Webpack to ignore some files from the "unused files" rule, so that's what I implemented here. I don't know if there is a standard way of doing that, I Googled but didn't find anything. My proposal is to prepend unused component files with an underscore, and then I added that ignore rule to Webpack configuration.