Remove jss-insertion-point
It seemed like we needed a manual jss insertion point to make the CDS Switch component layout work in RTL, but it appears that `body[dir='rtl']` styling in the component module is enough to make it work, and the insertion point is causing a weird bug with FlashMessage. Removing the insertion point introduces no regressions and removes the weird layout bug.