Relay Modern Upload
This commit adds a new Relay environment to handle uploading files in the Relay Modern style. `EnvironmentModern.js` exports a `createEnvironment` function. This is used in the graphql wrapper in `Newsletter` network logic that wraps `NewsletterComponent`, passing the relay modern `environment` for the update mutations to use. `me.token` and `team.slug` are required to create the evironment so these are also queried on `Newsletter` (but not the component, due to our relay linter rules keeping the concerns nicely separated!).

`EnvironmentModern.js` is basically a copy/paste of what we use in `check-mark`. The hard part was figuring out how to best provide a one-off environment for a specific graphql interaction -- it turns out that commitMutation in Relay Modern accepts an `environment` parameter so that was really all it took.

As of this commit, `Upload` is a placeholder component with more work to come.