UI for Upload component
`Upload` cares that it is being given a `fileName`, which as far as the component is concerned means there is a file that was returned by the API. It's also passed a `setFile` which it uses to set the file via the HTML drag and drop APIs (or a standard click to attach file), and also the `setFileName` which is used to clear out the `fileName` when a user unattaches a file.

There is logic in `NewsletterComponent` for a `fileName` state, whose initial state is derived from `header_file_url` -- but we no longer treat `header_file_url` as a malleable state since that's not settable in the API. (The only way to change the setting is to upload a new file.)

The Upload CSS uses CSS grid for layout in one place where SVG needs to be laid out in a certain way relative to labels.