Reload the page when newsletter is created for the first time
We are using a custom Relay network environment, since we need to handle file uploads. The problem is that the Relay environment is not updating the Relay store as a result of a mutation. We need to fix it properly in the Relay network layer. For now, this commit adds two workarounds for the newsletter UI:

* Reload the page when a newsletter is created
* Force a query refresh when changing language