Default newsletter timezone is local browser tz
When figuring out what value to display in the `NewsletterScheduler` we check for both `code` ('Europe/London') and `value` ('Europe/London (GMT+1:00)'). If the timezone comes from the browser's default timezone then it will match `code` and if it comes from the API it will match `value`. Doing it this way means we don't have to recalculate the offset in the parent `NewsletterComponent` or make changes to how we save the timezone in the API.

(#1439)