Disable newsletter scheduler for past datetimes
For static type newsletters we can't let users schedule in the past. This code takes the date, time, and time zone of the scheduler and calculates the unix epoch for that exact moment. Then it compares that epoch time to the current, in browser epoch time. If the scheduled time is in the past then we disable saving. We also display an error message informing the user.

#1441