Validate field length in newsletter settings
This commit adds an `onErrorTooLong` function prop to `LimitedTextArea` -- a function that is called, and passed a binary error state as its first and only parameter. This allows controlling forms to set state variables etc as needed when a LimitedTextArea's error state changes due to going over or under the character limit.