Clear errors on newsletter header file upload component when file is cleared.
In newsletter settings, if you choose a header of type Image (or another that does upload) and attach a file with the wrong file type (like PDF), you correctly see a message that says the file couldn’t be uploaded and needs to be of type png, jpg, or jpeg. But if you then attach a new file of type png/jpg/jpeg, the error message remains and says that the new file needs to be of the correct file type.

This fix clears out the error state when an attached file is removed or changed.

Fixes: CV2-3166.