Fix crash on trying to url match youtube homepage
There was an errbit crash where a regex was returning `null` where we didn't expect it. This was because the media URL recorded for the page was just `https://www.youtube.com`. This meant that `isYoutube` was true because of the domain match, but the regex for youtube videos doesn’t match because the regex is not checking for the youtube home page (nor should it). Here we treat something as a youtube video only if `isYoutube` is true and the file path matches what we expect from a youtube video.