Change how we detect YouTube videos
There was an errbit bug where we were parsing `https://www.youtube.com` as though it were a video. This is because we were checking against the domain `youtube.com` and assuming something must be a youtube video if that's the domain. But we can't assume that and it results in sometimes crashes and sometimes just strangely rendered media pages when a user puts in a youtube URL that isn't a video or a channel. So now we explicitly check whether something is a _video_ if that is what we are actually needing in an individual scenario.

#1478, CV2-3284