TagList component
I didn't catch until tonight that this tagging system is different from the existing one (on the Feed model rather than the media tags). The design for the component says "use the existing Tag selection menu component [aka TagMenu]" but that's not a generic UI component. Rather it's hard coded to work with media tags. So something that I thought would be "oh I'll just drop the existing TagMenu component in" is now more like I'm going to have to create a generic, non-data-bound version of the tag selection menu (which is good, we want that anyway). It will mostly be copy/pasting TagMenu and removing the fragment, instead relying on data passed in to the component, but it's an extra thing that I hadn't expected to come up.

This version is a version of TagList that renders a list of tags and has a hook for deleting them, but creating new tags will come in a future commit.