CV2-2830: Create shared feed UI page and refactored tagging components (#1476)
This commit adds a new Create Shared Feed page, and a button to navigate to it on the sidebar. Implements tooltips for all "create" (+) buttons on the sidebar.
Also adds a SelectList component that contains a QueryRenderer wrapping a regular Select, and new Tags component refactored and integrated by Darius to the new create shared feed page.