Fix crash when adding new item to blank media (imported report)
This commit adds a team fragment to MediaCardLarge so MediaStatusCommon doesn't crash when creating a new media.
It also skips the mutations optimisticResponse when performing this action. The optimisticResponse was another point of crash in
this operation but it is only needed when creating new items from the list view, so the fastest fix was to skip it.

References: CV2-2984