Fixing unit tests
When we switch from a React component (such as a material-ui icon) to a bare SVG (like our custom icons), we can no longer use Jest's "find the component" shortcut selector. Instead we need to apply a class name to the SVG and use a basic CSS selector for that.