Improve LimitedTextArea and TextField perf
`TextField` and `TextArea` were both rerendering their *parent* components on every keystroke. We normally solve this by switching from `onChange` to `onBlur` for setting values, but a problem arises with `LimitedTextArea`, which requires its own state to change on every keystroke (in order to display the number of remaining characters). If we did the normal `onBlur` solution then the character count would only update on blur.

`TextField` and `TextArea` now use the `forwardRef` pattern. This lets us set a `ref` on them just like we might do on an `input` element (by default in React, `ref` only works on basic HTML elements and not components -- using `forwardRef` allows us to use `ref` on components as well). Other than wrapping these in `forwardRef` and adding the forwarded `ref` to the `textarea` and `input` HTML elements, no changes were made to the functionality of these components. So they behave exactly how they did before, they just also have the ability to take a `ref`. Importantly this means we have at least 3 options when using `TextField` and `TextArea`:

 - use a standard onChange pattern where you pass `value={myText}` and `onChange={e => setMyText(e.target.value)}` (rerenders the parent every keystroke)
 - use an onBlur pattern instead of `value={myText}` and `onBlur={e => setMyText(e.target.value)}` (rerenders the parent on blur)
 - and, new as of this commit, declare `const inputRef = React.useRef()` in the parent component, pass `ref={inputRef}` on `TextField` or `TextArea`, and then refer to `inputRef.current.value` in the parent to get access to the value in the text component (NEVER rerenders the parent and in fact just behaves like an uncontrolled component). If a local state tied to the value of the input is needed (like displaying character count) you can avoid a parent rerender by creating a local copy of the ref value and updating it on keystrokes but only ever referring to it in the scope of the current component (so the current component rerenders but the parent doesn't, much more efficient). You can see how it's done in `LimitedTextArea.js` for an example of implementing this style of input.

Meanwhile, in addition to implementing the third solution above, `LimitedTextArea` gets a change in its `useEffect` for error handling. We do want to check for errors on every rerender of `LimitedTextArea` so the `useEffect` is correct here, but previously we were constantly updating the parent element's `onErrorTooLong` on every keystroke! So even with the ref stuff implemented we were *doubly* rerendering the parent. This commit makes a change where we only update the error state on the parent when we *transition* from one error state to another. So ultimately, what triggers a rerender of the parent of `LimitedTextArea` is: an onBlur event, or a change from an error state to a non-error state.

Other changes:

 - add `LimitedTextArea` to the UI sandbox
 - make `NewsletterStatic` update article values onBlur instead of onChange
 - updated the keys in `NewsletterStatic` to use index numbers and squashed the linter error - normally this is unadvised but in this case it's an array we only append items to or remove items from the end so it's fine, and there was a little hiccup where the keys were all `undefined` until the data loaded from the API so this squashes that warning