Deprecating next/previous media item nav
This removes the `NextPreviousLinks` component from the header in the media item view. Also removing the unit tests that check for calculating sibling nav urls when `MediaPage` is rendered, since that was only needed for the `NextPreviousLinks` component.

The component, along with the helper function `getListUrlQueryAndIndex`, is still used in the `media/FeedItem.js`

CV2-3361