Fix error in super admin screen logic
This is a tricky one -- localStorage only stores strings, so you might think that checking `superAdminMaskSession !== 'false'` would tell you if it was true, but in fact you need `superAdminMaskSession === 'true'` -- this is because in the case where `superAdminMask` is not set at all, you get `null`. So regular users who had never touched `superAdminMaskSession` in their local storage would get `null`, which does not equal `'false'`, so the code thought the screen should show because `'false' !== null`.

CV2-3228