Unmatched/matched media filter
This commit adds a `oneOption` prop on `MultiSelectFilter`. This lets us have a filter item that offers no further inputs to the user when clicked -- rather, we pass in a predefined set of selections (aka array of string) and assume that is the only correct state when this filter is selected. The commit also adds a new filter, "Media unmatched", to the filters list that uses this new option and sets `unmatched` to `['1']` in the search query.

CV2-3528, #1561