Ignore raised errors in integration tests with flag set
Sometimes we want to wait for an element to load and simply know if the element was there or not -- prior to this our helper functions would raise an error if the element didn't exist. This `ignoreRaise` flag is a way to optionally not raise an error and just return an empty array if nothing is there.