Fixing date parsing in French (and possibly other languages)
In the request card, when using the application in French, the date a report was received is not displayed correctly. It either shows "invalid date" or a date where month and day are swapped (for example, "August 10" in English shows as "October 8" in French).

This is because the `<RequestSubscription />` component passes its `lastCalledAt` `prop` as the `value` `prop` of a `<FormattedDate />` component, so it's expected to be a timestamp (string or number) or a date object, not a formatted string.

Possibly happens to other languages too - it doesn't happen with English though.

I implemented unit tests that were able to reproduce the issue. The tests passed after the fix.

I also took the opportunity to fix some `PropTypes` warnings raised by the unit tests.

Fixes: CV2-3594.