Improve developer experience when doing automated testing.
There are two things today that are painful in terms of developer experience when dealing with integration tests in Check Web:

1. If we want to disable test retries (in order to develop faster), we need to manually change a line in a file
2. If we want to run integration tests in a branch, we need to manually comment a bunch of lines across two different files

These two things require work, are fragile, pollute the commit history for the branch and we can easily end up pushing these changes to the main branches by mistake.

This PR tries to address this in two ways:

1. Run integration tests when the commit message contains `[full ci]`
2. Control number of test retries using an environment variable `TEST_RETRY_COUNT`

Documentation (in the README) updated accordingly.

Reference: CV2-3669.