Fixing linter errors
There is a really silly interaction between the `dataset` DOM API and a linter rule that is 99% of the time the correct thing to do. Basically: when you have a function that is passed `event` and you set the data attribute using `event.target.node.dataset = blah`, that is correct and works! But the linter rule doesn't want you writing over any arguments passed in a function. The solution here is to use the alternative way of setting the attribute: calling `setAttribute` on the node.

This is really just linter pedantry clashing with a convenient but non-idiomatic DOM API feature.