Alternative approach to an autogrow textarea
Since the previous solution was a bit buggy in certain edge cases, we have changed tactics. The solution used here is to surround the `textarea` with a `div` which grows and shrinks due to being rendered as a `display: grid`. The `textarea` fills the `div`, and minimum height is now handled by setting `rows` rather than a height in pixels.

To set a maximum height, you can just pass in a simple custom style object on a case-by-case basis. There is now an example in `Sandbox`.

CV2-3180, #1624 

---------

Co-authored-by: Darius Kazemi <darius@meedan.com>