Hotfix: initial height of pre-populated TextArea
We need to set the height of TextArea to the correct value if it has text already in it. What I do here is repeat what we normally do to get the correct height but make sure to do it on initial render using the `forwardRef` to access the element instead of via `event`.