Chat history view (#1666)
pagination is implemented but not enabled -- this is due to two pagination UX issues. One is that there is not currently a way to go to a specific message via the API without first grabbing the array of all messages, which defeats the point of pagination. Still, we could paginate by loading just the most recent ~10 messages and going back from there. But the problem here is the second issue: delivery confirmation messages are their own message -- it's possible we paginate and there's a boundary in between a message and its delivery confirmation, which is just weird. What do we do there? For now I am not paginating while we address these issues, but the components are in there.
---------

Co-authored-by: Caio <117518+caiosba@users.noreply.github.com>