Forward a ref if missing
Basically what this code does is check to see if there is already a ref passed in to the TextArea component. If there is a ref, then it forwards that on to TextField. If there is not a ref, then it creates one and passes it on. Either way it means that TextField has a ref to a parent component where it can set the data attribute which the CSS `content` field can expand to.