Textarea Improvements CV2-3849 (#1678)
* set up textarea and sandbox for easier interaction testing CV2-3849

* Forward a ref if missing

Basically what this code does is check to see if there is already a ref passed in to the TextArea component. If there is a ref, then it forwards that on to TextField. If there is not a ref, then it creates one and passes it on. Either way it means that TextField has a ref to a parent component where it can set the data attribute which the CSS `content` field can expand to.

* correctly get the same height on the psuedo element as the text area to get the overall height correct CV2-3849

* better alignment so that textareas and textfield text are in the same visual location CV2-3849

* final tweak for heights and scrollbars to match correctly CV2-3849

---------

Co-authored-by: Darius Kazemi <darius@meedan.com>