Fix tags column crashing on initial load
When we change the column settings so that the Tags column is in there, and then navigate to All Items, there is an initial render of the component where `tags_as_sentence` is undefined that throws an error, followed by the actual updated render. This commit briefly renders an empty tags column until the the data is available shortly after.

Also a minor CSS fix to how the tags are displayed.