UI component to manage NLU keywords.
Super-admins see the tags component on the dialog that opens when they try to edit a tipline menu option. They can use that component to add and remove keywords for the NLU feature.

When a tipline menu option is associated with a keyword or phrase, any message sent to the tipline that matches that keyword will have the same effect as if that menu option was clicked.

Reference: CV2-3709.