Filter shared feeds by organization
This implements a series of filters available at the top of a Shared Feed view where the user can toggle on/off fact-checks from different organizations in the shared feed. There is also a Quick Filter that lets you pick your own org, or all orgs that are not your own. I'll point out the `toHaveBeenCalledWith` in the unit test -- it's a really handy function for testing whether a mock function was called with certain arguments.

#1707, CV2-3611