In edge cases where `media` is null, render error on SmallMediaCard
This usually happens rarely in production, maybe once a month across all users. It's probably due to a network error and can be resolved with a reload. There is a display issue in MediaSuggestionsComponent where the accept/reject/spam/trash buttons shouldn't be shown, but that has nothing to do with SmallMediaCard so I'm controlling it on a special case basis from MediaSuggestionsComponent.