Address full screen element via unique identifier
We had been picking the first instance of `.video-media-card` to fullscreen, but this fails when there is more than one `.video-media-card` in the DOM. Also, it meant that this component had a dependency outside of itself (namely it needed an element with that className to be present on the DOM to actually work).

Here we add a query for `id` (which is always a valid classname) and apply it to the class of the component itself, then we just go up the DOM tree and fullscreen whatever the parent of this component is. This means it'll work to fullscreen anything it's a direct child of, not just `.video-media-card`.

CV2-4036