Editors should not be able to manage feed collaborators
I had to wrap the `<SaveFeed />` component in a new `<CreateFeed />` component, which is basically a query renderer, and that now responds to the `feed/create` route. This way I'm able to get the permissions.

The rest of the code is basically checking for permissions and then rendering components or not based on them.

Reference: CV2-3953.