Safe property access/conversion
The original code could be the source of two errors -- one, if `list_columns_values` is undefined an error would throw for the unsafe property access. Another is if `published_by` is undefined (or even just not an object), the `Object.values` call will throw an error because it can't extract values from anything but an object. From a user perspective, the rest of the file will behave correctly because if published_by is undefined, then the length of `names` will be 0 and will render to the user as "-", which is what we'd want.