Design updates for feed fact-check card
Applying a few design updates for the fact-check card displayed on the feed page. Some of those changes were implemented in the more generic Check Design System component `<Card />`. I expect further refactoring to go into that component in the scope of CV2-3991 and CV2-3922. I'm going to talk more about it with @dariusk.

Summary of changes to fact-check card:

* Fact-check URL with icon, which is truncated when too long
* Workspace avatar at the bottom
* Small tweaks to the status/rating top-right label
* Fixing a couple of CSS bugs that exist on `develop`
* The whole card is not clickable anymore

Reference: CV2-3990.