Item Card refactor and bugfix
As preliminary work for CV2-3991, I've broken the `Card` component into three subcomponents: one for displaying a date, one for displaying the collapsible title/description, and one for displaying the item rating. The idea here is to eventually turn `Card` into just a thing that you hover over and it gets a highlight and also can double as a clickable link card if you pass it a url, but otherwise to be content-agnostic. Then we can have the layout for fact checks defined in `FactCheckCard`, the layout for shared feed item cards in a `SharedFeedItemCard`, etc etc. So ultimately our display cards will look like:

 * an overall layout logic component like `SharedFeedItemCard`
   * `Card`: basically just the rounded box container with hover and link behavior
     * a bunch of components we can pick and choose that actually do things in the layout like `ItemRating`

This also contains some bug fixes to behavior for the collapsible description fields: for example, it fixes an issue where long titles remained collapsed when a user pressed the expand button. There are also minor layout changes to make things more consistent with the CDS design (making the rating icon 16px instead of 14px, fixing some font settings, using `gap` instead of a space character to define spacing, etc).