Fix double-preview on youtube videos
We were showing both the embedded youtube player and also the web page preview for the video, due to the fact that youtube pages are also web pages. This was confusing since it looked like a double-video. This fix only shows the youtube preview for youtube videos, removing the double effect.