Refactor Card and FactCheckCard
I'm considering `Card` to be an empty rectangle container that can be passed a `cardUrl` and a `footer`. The card manages an `isHovered` state that is provided to its `children` via `Context` as needed. So in this case `FactCheckCard` renders a `Card` with some children in the correct layout. The `ItemDescription` is wrapped with the `CardHoverContext.Consumer`, because `ItemDescription` needs to know that `Card` (not `FactCheckCard`!) is being hovered. In other cases I would do this by passing state from Card to its children, but since we have a generic `children` object I'm doing it this way instead. Seems to work well.