New feature: Custom item title
This PR adds a new feature: custom item title. On the item page, the header now shows a component where the title for an item can be customized. The user is able to type a custom title or choose between claim title, fact-check title and pinned media ID. See referenced ticket for details on designs and specs.

In terms of coding, most of the work here happens in a new component `<ItemTitle />`. Very little was changed in existing code.

I used states and not only props on this new component because there were situations where the component was not re-rendering (I think because props were changed from inside and so not propagating). I thought it was simpler to just be sure that the component had its own state.

I noticed that in recent components, like the feed quick filters, we're still using Material UI menu components, so that's what I used here as well.

Reference: CV2-3986.