Conditional media icon when count is 1
The icon on the media count shows the correct icon type if the count is 1. The component does require its parent to pass in the `mediaType` when necessary since a display component shouldn’t be deducing that on its own. If the count is 1 but `mediaType` is not defined, then it defaults to the generic media icon.