Refactor Card and FactCheckCard, implement SharedItemCard
* Item Card refactor and bugfix

I've broken the `Card` component into three subcomponents: one for displaying a date, one for displaying the collapsible title/description, and one for displaying the item rating. The idea here is to eventually turn `Card` into just a thing that you hover over and it gets a highlight and also can double as a clickable link card if you pass it a url, but otherwise to be content-agnostic. Then we can have the layout for fact checks defined in `FactCheckCard`, the layout for shared feed item cards in a `SharedItemCard`, etc etc. So ultimately our display cards will look like:

 * an overall layout logic component like `SharedItemCard`
   * `Card`: basically just the rounded box container with hover and link behavior
     * a bunch of components we can pick and choose that actually do things in the layout like `ItemRating`

This also contains some bug fixes to behavior for the collapsible description fields: for example, it fixes an issue where long titles remained collapsed when a user pressed the expand button. There are also minor layout changes to make things more consistent with the CDS design (making the rating icon 16px instead of 14px, fixing some font settings, using `gap` instead of a space character to define spacing, etc).

I'm considering `Card` to be an empty rectangle container that can be passed a `cardUrl` and a `footer`. The card manages an `isHovered` state that is provided to its `children` via `Context` as needed. So in this case `FactCheckCard` renders a `Card` with some children in the correct layout. The `ItemDescription` is wrapped with the `CardHoverContext.Consumer`, because `ItemDescription` needs to know that `Card` (not `FactCheckCard`!) is being hovered. In other cases I would do this by passing state from Card to its children, but since we have a generic `children` object I'm doing it this way instead. Seems to work well.

The ItemChannels subcomponent accepts an object formatted like `{ main: 8, others: [7, 8, 9, 10] }` as expected from the API, then looks up channel values in `CheckChannels` and uses a custom human-readable channel name I pulled from `TiplineRequest` for the tooltip values.

There's a `sortMainFirst` prop that is to be `true` for `SharedItemCard` but not for a general workspace item card. Setting this to true means that if the API indicates the first tipline an item was reported on, that can be first in the list.

The icon on the media count shows the correct icon type if the count is 1. The component does require its parent to pass in the `mediaType` when necessary since a display component shouldn’t be deducing that on its own. If the count is 1 but `mediaType` is not defined, then it defaults to the generic media icon.

`ItemDate` takes a `Date` object instead of a unix timestamp, since these components shouldn't be responsible for conversions and time zones etc; we should simply be passing a correct `Date` object into their props and counting on the timezone data provided by the API to be correct.

Includes fixes for layout with long fact check urls (the key was wrapping the url in a div and setting the div to `display: grid`!), 

Include CheckFeedDataPoints, card variants based data points

---------

Co-authored-by: Caio <117518+caiosba@users.noreply.github.com>

CV2-3991, #1785