Manage languages with useState in LanguagesComponent
Adding and removing languages to the list was not working in certain cases -- this is because languages weren't represented as state on `LanguagesComponent`. By changing it to a `useState` variable and passing `setLanguages` to its children doing the mutations, we trigger a rerender on `LanguagesComponent` when langs are added or removed.