Fix "sticky" tooltip
This was a bug introduced in 95ca6ef12f -- we added `disableHoverListener`, which not only disables hovering, but disables the listener that tells when hovering *finishes*. So if you press "next" until the button is disabled, then mouse away from the button, the tooltip does not get removed, because the hover listener is gone. Removing this entirely, which means the tooltip will appear on the disabled button, but better than having it stick around forever.