Added an alert when a user tries to load beyond result 10000
Elastic Search is limited to 10k results returned, and after that it fails to load more.
It's an edge case because paging all the way to 10k is a long process, so a simple error saying to narrow down the search results is sufficient.
Added an Alert component that opens where the user is clicking and tells them to filter their results
Removed the Link component and used onClick to check the high end of the results range before loading the next page

CV2-4312