Shared feed item page (#1849)
Now clusters (listed in the feed page for a feed that is sharing media) are clickable and take to the feed item page. The feed item page includes a header, a list of workspaces and their claims/fact-checks and a list of media. Both fact-checks and media can be clicked and have details displayed in a modal.

References: CV2-4175, CV2-4176, CV2-4177 and CV2-4178.