Feeds created in different timezones will sometimes show different numbers of items.  To limit some confusing, it was decided that the UTC offset should be seen next to date to indicate the actual range
If an ISO string is converted back to a date, it assumes the local timezone unless one is stated, so using Date's getTimezoneOffset wasn't an option.
The input field of the DatePicker requires an ISO value to display the correct date, so that couldn't be changed.  Plus, all saved list ranges are currently saved as ISO strings.
Knowing that all start dates are made at 00:00:00 and end dates are 23:59:59, we can work backwards to figure out the original timezone the string was created in.
Countries like India also have a UTC offset of +5:30 so minutes have to be considered.

CV2-4416 - Applying the same filter displays different number of items