Localization refactor
This commit changes the way we are handling labels to something more react-intl appropriate. Specifically we use `{variableName, number}` in the template string to get the equivalent of doing `Intl.NumberFormat(locale, {})` on the number itself. (We can't do the formatting before passing it in because we get `NaN` due to the fact that it's a string with commas etc.)

This also refactors `SharedItemCardFooter` to use the pre-existing `MediaCount`, `SuggestionsCount`, etc. It also changes `SharedItemCardFooter` to include items when e.g. `mediaCount !== null` -- if we are just looking for truthy values (`mediaCount &&`) then we will never display when there are 0 requests etc.