Workspace Item Card (#1808)
* MediaCount component

* RequestsCount component

* LastRequestDate component

* Small fix - add rating/color props to SearchResultsCards

* Further WorkspaceItemCard work

* There is now a `WorkspaceItemCard` that uses `Card`
* The `WorkspaceItemCard` currently renders `ItemDescription` and `ItemThumbnail`
* I've added an unread indicator controlled by an `isUnread` prop on the component.
* I've added an "empty state" to `ItemThumbnail`. @caiosba Please take a close look at `ItemThumbnail`'s `propTypes` -- I implemented some custom prop type functions. Basically, it means that if we specify a `type` then it expects all fields to be required, but if `type` is empty, then no fields are required and it returns the new "empty state" icon seen in the screenshot above. I'm not sure we use custom proptypes anywhere else in the code so I wanted to flag that.
* Also I fixed an issue with `SharedItemCard` (in its own commit), which is

* Adding multiple props and display widgets to WorkspaceItemCard

Also updating the PropTypes on MediaCount

* Adding `isPublished` and `rating` features

The `ItemRating` component now displays a "no fact-check" message if it is not passed a `rating`.

* Adding localized strings for ItemRating

* Add SuggestionsCount component

* Checkboxes for WorkspaceItemCard

If the component is provided `onCheckboxChange`, then a checkbox is rendered.

* Final checkbox functionality, trash implementation

Had to add some click swallowing code to the `Checkbox` component in order to not trigger the click on the `Card` when clicking a checkbox.

* Adding rating color

* Using browser locale instead of en-US

* Localization refactor

This commit changes the way we are handling labels to something more react-intl appropriate. Specifically we use `{variableName, number}` in the template string to get the equivalent of doing `Intl.NumberFormat(locale, {})` on the number itself. (We can't do the formatting before passing it in because we get `NaN` due to the fact that it's a string with commas etc.)

This also refactors `SharedItemCardFooter` to use the pre-existing `MediaCount`, `SuggestionsCount`, etc. It also changes `SharedItemCardFooter` to include items when e.g. `mediaCount !== null` -- if we are just looking for truthy values (`mediaCount &&`) then we will never display when there are 0 requests etc.

* Pass mediaType in for trash view; use thumbnail type as backup