Added NavigateAwayDialog to SaveFeed, preventing user accidentally losing data when leaving the page.
The unique ways each piece of data gets updated on this page made a single check for any update a little tricky.  I used a useEffect hook that checks for changes to each of the data components to avoid making changes to every onChange. A successful save switches the flag to false, allowing navigation away.  Tested by adding and removing contributors, changing name, removing List.

Adding a check during editing for if the updated value matches the original value felt like overkill, but there is a case where a user undoes their change so it appears nothing is different but the input still went through changes so the popup still appears.

Ticket: CV2-4060