Change to the NavigateAway solution
Refactored the inputs into a single state object that uses a single update function.  This is a change from each field having its own state and update function.  The reason is to consolidate the updates into a single location making it easy to tell when the form is dirty. Three additional handlers are used for updating values that use external components, to avoid changing any other files.  This also makes it easier to see what data can be changed in the form.