Fix/cv2-4026: shared feed invites should not show up in the workspaces where they were created (#1891)
* remove duplicate 'FeedInvitation' feeds based on their 'dbid' property

For each feed in the feeds array, it checks if the feed type is 'FeedInvitation'. If so, it uses the some method to determine if any other feed in the array has the same 'dbid'. If a duplicate 'FeedInvitation' feed is found, it is excluded from the filteredFeeds array

Reference: CV2-4026