Cv2 4060 add navigate away dialog to shared feeds (#1892)
* Added NavigateAwayDialog to SaveFeed, preventing user accidentally losing data when leaving the page.

The unique ways each piece of data gets updated on this page made a single check for any update a little tricky.  I used a useEffect hook that checks for changes to each of the data components to avoid making changes to every onChange. A successful save switches the flag to false, allowing navigation away.  Tested by adding and removing contributors, changing name, removing List.

Adding a check during editing for if the updated value matches the original value felt like overkill, but there is a case where a user undoes their change so it appears nothing is different but the input still went through changes so the popup still appears.

Ticket: CV2-4060

* Change to the NavigateAway solution

Refactored the inputs into a single state object that uses a single update function.  This is a change from each field having its own state and update function.  The reason is to consolidate the updates into a single location making it easy to tell when the form is dirty. Three additional handlers are used for updating values that use external components, to avoid changing any other files.  This also makes it easier to see what data can be changed in the form.

* Updated the Formatted message from Tasks to SaveFeed

* added localization file for task -> saveFeed change

* fix linter errors after merging

* Fix merge error

* Fix SaveFeed not navigating to Feed page

---------

Co-authored-by: Alexandre Amorim <amoedoamorim@gmail.com>